# server/openapi.yaml
openapi: 3.0.3
info:
  title: Valance API
  version: 2.0.0
  description: |
    Social habit-tracking backend. Base URL: http://localhost:3000/api/v1

    **Auth:** Bearer token (Firebase JWT) in Authorization header. Dev mode: X-Dev-User-Id header bypasses auth.

    **Response format:** `{ status: "ok", data: ... }` or `{ status: "error", error: { code, message } }`

    **Dates:** YYYY-MM-DD strings. Timestamps: ISO 8601.

    **Constants:**
    - XP/Sparks: light=5, moderate=10, intense=20, perfect_day=25, welcome=50
    - Streak milestones: 7d=+30, 30d=+100, 100d=+500
    - Ranks: bronze(0), silver(500), gold(2000), platinum(5000), diamond(15000)
    - Group tiers: spark(0), ember(7), flame(21), blaze(66)
    - Group chain: 100%=gold, ≥75%=silver, <75%=broken
    - Max habits: 7 active per user
    - Freeze: 100 Sparks, 1/group/day
    - Nudge limit: 5/day, 1/pair/day
servers:
  - url: http://localhost:3000/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List user's groups
      description: Returns all groups the authenticated user belongs to, with their role, member count, tier, and streak info.
      responses:
        '200':
          description: Array of groups
          content:
            application/json:
              example:
                status: ok
                data:
                  - id: "9b21f6ae-..."
                    name: "Morning Warriors"
                    inviteCode: "j8gjVF_H"
                    tier: "ember"
                    currentStreak: 12
                    longestStreak: 14
                    totalLinks: 18
                    role: "admin"
                    joinedAt: "2026-03-15T10:00:00Z"
                    memberCount: 5

    post:
      tags: [Groups]
      summary: Create group
      description: Creates group with auto-generated invite code (8 chars). Creator becomes admin. Size 3-7 members.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
            example:
              name: "Morning Warriors"
      responses:
        '201':
          description: Group created with invite code
          content:
            application/json:
              example:
                status: ok
                data:
                  group:
                    id: "9b21f6ae-..."
                    name: "Morning Warriors"
                    inviteCode: "j8gjVF_H"
                    tier: "spark"
                    currentStreak: 0
                  membership:
                    role: "admin"
  /groups/{id}:
    get:
      tags: [Groups]
      summary: Get group detail
      description: Returns group info, today's member completion status (respects minimal visibility), chain data.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Group detail
        '403':
          description: Not a member
  /groups/{id}/join:
    post:
      tags: [Groups]
      summary: Join group
      description: Join via invite code. Max 7 members.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code: { type: string }
      responses:
        '201':
          description: Joined
        '400':
          description: Wrong invite code or group full
  /groups/{id}/leave:
    delete:
      tags: [Groups]
      summary: Leave group
      description: If admin leaves, transfers admin to next oldest member.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Left group
  /groups/{id}/feed:
    get:
      tags: [Groups]
      summary: Group activity feed
      description: Completions, kudos, nudges, chain links, tier changes, Status+Norm messages. Chronological.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Feed items with actor info
  /groups/{id}/streak:
    get:
      tags: [Groups]
      summary: Chain link history
      description: Last 30 days of gold/silver/broken links. Current streak, longest, total links.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Chain link data
  /groups/{id}/leaderboard:
    get:
      tags: [Groups]
      summary: Intra-group leaderboard
      description: Weekly contribution score = habits completed + gold link contributions + kudos received. Resets every Monday.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: period
          in: query
          schema: { type: string, enum: [week, month], default: week }
      responses:
        '200':
          description: Ranked members
  /groups/{id}/members:
    get:
      tags: [Groups]
      summary: List group members
      description: All members with today-completion status. Respects minimal visibility (only done/not done, no habit names).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Member list
  /groups/{id}/freeze:
    post:
      tags: [Groups]
      summary: Altruistic streak freeze
      description: Costs 100 Sparks from requesting user. Protects group chain for today. Max 1 per group per day.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Freeze applied
        '400':
          description: Insufficient Sparks or already frozen today
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT token. Dev mode alternative — X-Dev-User-Id header.
...